+++ ingeniamotion/motion_controller.py
@@ -10,6 +10,7 @@ from ingeniamotion.configuration import Configuration
 from ingeniamotion.drive_tests import DriveTests
 from ingeniamotion.errors import Errors
 from ingeniamotion.information import Information
+from ingeniamotion.io import InputsOutputs
 from ingeniamotion.metaclass import DEFAULT_AXIS, DEFAULT_SERVO
 from ingeniamotion.motion import Motion
 from ingeniamotion.fsoe import FSoEMaster
@@ -29,6 +30,7 @@ class MotionController:
         self.__tests: DriveTests = DriveTests(self)
         self.__errors: Errors = Errors(self)
         self.__info: Information = Information(self)
+        self.__io = InputsOutputs(self)
         self.__fsoe: FSoEMaster = FSoEMaster(self)
 
     def servo_name(self, servo: str = DEFAULT_SERVO) -> str:
@@ -145,3 +147,8 @@ class MotionController:
     def fsoe(self) -> "FSoEMaster":
         """Instance of :class:`~ingeniamotion.fsoe.FSoEMaster` class"""
         return self.__fsoe
+
+    @property
+    def io(self) -> InputsOutputs:
+        """Instance of :class:`~ingeniamotion.io.InputsOutputs` class"""
+        return self.__io